Rename some of the core file operations
