+++ packages/core/src/create/PublisherCore.ts
@@ -2,8 +2,8 @@ import type { TransactionBuilder } from "@metaplex-foundation/js";
 import debugModule from "debug";
 import type { Signer } from "@solana/web3.js";
 
-import { validatePublisher } from "../validate/index.js";
-import { Constants, mintNft } from "../utils.js";
+import { validatePublisher } from "../validate/CoreValidation";
+import { Constants, mintNft } from "../CoreUtils";
 import type { Context, Publisher, PublisherMetadata } from "../types.js";
 
 const debug = debugModule("PUBLISHER");
