+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./create/index.js";
 export * from "./publish/index.js";
-export * from "./validate/index.js";
+export * from "./validate/CoreValidation";
 export * from "./types.js";
